backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
fairlearn>=0.9.0
google-generativeai>=0.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
reportlab>=4.0.0
python-dotenv==1.0.1
pydantic[email]>=2.5.0
aiofiles>=23.0.0
